% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-function.R
\name{extract_infos_from_names}
\alias{extract_infos_from_names}
\title{Extract infos from barcode}
\usage{
extract_infos_from_names(barcodes)
}
\arguments{
\item{barcodes}{a vector of barcode}
}
\value{
data.frame
}
\description{
Extract infos from barcode
}
