% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-function.R
\name{set_expand}
\alias{set_expand}
\title{Expand collapsed "cols" aìin a DataFrame df}
\usage{
set_expand(df, cols, sep = ";", pairs = TRUE)
}
\arguments{
\item{df}{a DataFrame}

\item{cols}{the colums that need to be expanded}

\item{sep}{separator character}

\item{pairs}{logical; split as pairs}
}
\value{
DataFrame
}
\description{
Expand collapsed "cols" aìin a DataFrame df
}
