% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-function.R
\name{import_table}
\alias{import_table}
\title{Wrapper for read table}
\usage{
import_table(file, header = T, row.names = 1)
}
\arguments{
\item{file}{a filename}

\item{header}{if TRUE consider first line header}

\item{row.names}{assume column as row.names}
}
\value{
data.frame
}
\description{
Wrapper for read table
}
