% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-functions.R
\name{imputeKNN}
\alias{imputeKNN}
\title{Impute values}
\usage{
imputeKNN(...)
}
\arguments{
\item{\dots}{all arguments are passed to impute.knn}
}
\description{
Wrapper to impute.knn. ?inpute.knn to see arguments.
}
