% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations-handler.R
\name{prepareMutations}
\alias{prepareMutations}
\alias{summarizeMutation}
\alias{selectMutation}
\alias{selectImpact}
\title{From maf format prepare the mutation table}
\usage{
prepareMutations(maf, impact = NULL, mutType = NULL,
  filterByThisEntrez = NULL, patients = NULL, select_type = TRUE,
  tumor_site_type = "01", select_columns = c(hugo = 1, entrez = 2,
  impact = 94, type = 9, patient = 16))

summarizeMutation(smallMaf)

selectMutation(smallMaf, type)

selectImpact(smallMaf, type)
}
\arguments{
\item{maf}{a maf format table}

\item{impact}{select by impact}

\item{mutType}{select by mutation type}

\item{filterByThisEntrez}{filter to keep only entred listed}

\item{patients}{keep only the patients listed}

\item{select_type}{logical to perform selection on the type}

\item{tumor_site_type}{keep only tumor samples of the type. 01 primary solid tumor.}

\item{select_columns}{named numeric to specyfy the columns to select}

\item{smallMaf}{reduced version of a maf file}

\item{type}{mutation type}
}
\value{
a matrix entrez x samples

smallMaf subset (matrix or data.frame)
}
\description{
From maf format prepare the mutation table

Collapse mutations
  Given the same gene and patients mutation impact and type are collapsed.

Select mutation by type

Select mutation by impact
}
